# IotFunctions/IotApis/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "IotApis",
      "version": "1.0"
    },
    "paths": {
      "/api/camera/devices/{deviceId}/imagedata": {
        "get": {
          "tags": [
            "DeviceImageData"
          ],
          "operationId": "GetDeviceImageData",
          "parameters": [
            {
              "name": "deviceId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "dateFrom",
              "in": "query",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "dateTo",
              "in": "query",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "preset",
              "in": "query",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/DeviceImage"
                    }
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/DeviceImage"
                    }
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/DeviceImage"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/camera/devices/iotcentral/{deviceId}/imagedata": {
        "get": {
          "tags": [
            "DeviceImageData"
          ],
          "operationId": "GetIotCentralDeviceImageData",
          "parameters": [
            {
              "name": "deviceId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "dateFrom",
              "in": "query",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "dateTo",
              "in": "query",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "authorization",
              "in": "header",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/camera/devices/iotcentral/{deviceId}/images/latest": {
        "get": {
          "tags": [
            "DeviceImageData"
          ],
          "operationId": "GetLatestIotCentralDeviceImageData",
          "parameters": [
            {
              "name": "deviceId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "authorization",
              "in": "header",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/camera/devices/iotcentral/{deviceId}/property": {
        "get": {
          "tags": [
            "DeviceImageData"
          ],
          "parameters": [
            {
              "name": "deviceId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "authorization",
              "in": "header",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/weather/devices/{deviceId}/telemetries": {
        "get": {
          "tags": [
            "DeviceTelemetry"
          ],
          "operationId": "GetDeviceTelemetries",
          "parameters": [
            {
              "name": "deviceId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "dateFrom",
              "in": "query",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "dateTo",
              "in": "query",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/DeviceTelemetry"
                    }
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/DeviceTelemetry"
                    }
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/DeviceTelemetry"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/weather/devices/iotcentral/{deviceId}/telemetries": {
        "get": {
          "tags": [
            "DeviceTelemetry"
          ],
          "operationId": "GetIotCentralDeviceTelemetries",
          "parameters": [
            {
              "name": "deviceId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "dateFrom",
              "in": "query",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "dateTo",
              "in": "query",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "authorization",
              "in": "header",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/weather/devices/iotcentral/{deviceId}/telemetries/latest": {
        "get": {
          "tags": [
            "DeviceTelemetry"
          ],
          "operationId": "GetIotCentralDeviceLatestTelemetries",
          "parameters": [
            {
              "name": "deviceId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "authorization",
              "in": "header",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/weather/devices/iotcentral/{deviceId}/property": {
        "get": {
          "tags": [
            "DeviceTelemetry"
          ],
          "parameters": [
            {
              "name": "deviceId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "authorization",
              "in": "header",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/images": {
        "get": {
          "tags": [
            "Image"
          ],
          "parameters": [
            {
              "name": "imagePath",
              "in": "query",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "DeviceImage": {
          "type": "object",
          "properties": {
            "deviceId": {
              "type": "string",
              "nullable": true
            },
            "id": {
              "type": "string",
              "nullable": true
            },
            "timeStamp": {
              "type": "integer",
              "format": "int64"
            },
            "cameraName": {
              "type": "string",
              "nullable": true
            },
            "cameraAddress": {
              "type": "string",
              "nullable": true
            },
            "snapshotDate": {
              "type": "string",
              "nullable": true
            },
            "partitionKey": {
              "type": "string",
              "nullable": true
            },
            "imagePath": {
              "type": "string",
              "nullable": true
            },
            "cameraModel": {
              "type": "string",
              "nullable": true
            },
            "cameraSerialNumber": {
              "type": "string",
              "nullable": true
            },
            "cameraManufacture": {
              "type": "string",
              "nullable": true
            },
            "preSet": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
        "DeviceTelemetry": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "nullable": true
            },
            "deviceId": {
              "type": "string",
              "nullable": true
            },
            "measurements": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              },
              "nullable": true
            },
            "timestamp": {
              "type": "integer",
              "format": "int64"
            }
          },
          "additionalProperties": false
        }
      }
    }
  }